Fix: Class "Inbox" has 26 methods, which is greater than 20 authorized. Split it into smaller classes.
